+++ common/transformations/model.py
@@ -1,8 +1,8 @@
 import numpy as np
 
-from common.transformations.camera import eon_focal_length, \
-        vp_from_ke, get_view_frame_from_road_frame, \
-        FULL_FRAME_SIZE
+from common.transformations.camera import (FULL_FRAME_SIZE, eon_focal_length,
+                                           get_view_frame_from_road_frame,
+                                           vp_from_ke)
 
 # segnet