+++ selfdrive/test/test_car_models.py
@@ -14,6 +14,7 @@ from selfdrive.services import service_list
 import selfdrive.messaging as messaging
 from common.params import Params
 from common.basedir import BASEDIR
+from common.fingerprints import all_known_cars
 from selfdrive.car.honda.values import CAR as HONDA
 from selfdrive.car.toyota.values import CAR as TOYOTA
 from selfdrive.car.gm.values import CAR as GM
@@ -54,6 +55,7 @@ def get_route_logs(route_name):
         with open(log_path, "w") as f:
           f.write(r.content)
       else:
+        print "failed to download test log %s" % route_name
         sys.exit(-1)
 
 routes = {
@@ -138,6 +140,10 @@ routes = {
   },
   "232585b7784c1af4|2019-04-08--14-12-14": {
     'carFingerprint': HONDA.CRV_HYBRID,
+    'enableCamera': True,
+  },
+  "99e3eaed7396619e|2019-08-13--15-07-03": {
+    'carFingerprint': HONDA.FIT,
     'enableCamera': True,
   },
   "2ac95059f70d76eb|2018-02-05--15-03-29": {
@@ -351,6 +357,21 @@ routes = {
     'enableCamera': False,
     'enableDsu': False,
   },
+  "2e07163a1ba9a780|2019-08-25--13-15-13": {
+    'carFingerprint': TOYOTA.LEXUS_IS,
+    'enableCamera': True,
+    'enableDsu': False,
+  },
+  "2e07163a1ba9a780|2019-08-29--09-35-42": {
+    'carFingerprint': TOYOTA.LEXUS_IS,
+    'enableCamera': False,
+    'enableDsu': False,
+  },
+  "1dd19ceed0ee2b48|2018-12-22--17-36-49": {
+    'carFingerprint': TOYOTA.LEXUS_IS, # 300 hybrid
+    'enableCamera': True,
+    'enableDsu': False,
+  },
   "791340bc01ed993d|2019-03-10--16-28-08": {
     'carFingerprint': SUBARU.IMPREZA,
     'enableCamera': True,
@@ -370,32 +391,58 @@ passive_routes = [
   #"bfa17080b080f3ec|2018-06-28--23-27-47",
 ]
 
-public_routes = [
-  "f1b4c567731f4a1b|2018-06-06--14-43-46",
-  "f1b4c567731f4a1b|2018-04-18--11-29-37",
-  "f1b4c567731f4a1b|2018-04-18--11-29-37",
-  "7ed9cdf8d0c5f43e|2018-05-17--09-31-36",
-  "38bfd238edecbcd7|2018-08-22--09-45-44",
-  "38bfd238edecbcd7|2018-08-29--22-02-15",
-  "b0f5a01cf604185c|2018-01-26--00-54-32",
-  "b0f5a01cf604185c|2018-01-26--10-54-38",
-  "b0f5a01cf604185c|2018-01-26--10-59-31",
-  "56fb1c86a9a86404|2017-11-10--10-18-43",
-  "b0f5a01cf604185c|2017-12-18--20-32-32",
-  "b0c9d2329ad1606b|2019-04-02--13-24-43",
-  "791340bc01ed993d|2019-03-10--16-28-08",
+# TODO: replace all these with public routes
+# TODO: add routes for untested cars: HONDA ACCORD 2018 HYBRID TOURING and CHRYSLER PACIFICA 2018
+non_public_routes = [
+  "0607d2516fc2148f|2019-02-13--23-03-16", # CHRYSLER PACIFICA HYBRID 2019
+  "3e9592a1c78a3d63|2018-02-08--20-28-24", # HONDA PILOT 2017 TOURING
+  "aa20e335f61ba898|2019-02-05--16-59-04", # BUICK REGAL ESSENCE 2018
+  "1851183c395ef471|2018-05-31--18-07-21", # HONDA CR-V 2017 EX
+  "9d5fb4f0baa1b3e1|2019-01-14--17-45-59", # KIA SORENTO GT LINE 2018
+  "b4c18bf13d5955da|2018-07-29--13-39-46", # TOYOTA C-HR HYBRID 2018
+  "5a2cfe4bb362af9e|2018-02-02--23-41-07", # ACURA RDX 2018 ACURAWATCH PLUS
+  "362d23d4d5bea2fa|2018-08-10--13-31-40", # TOYOTA HIGHLANDER HYBRID 2018
+  "aa20e335f61ba898|2018-12-17--21-10-37", # BUICK REGAL ESSENCE 2018
+  "215cd70e9c349266|2018-11-25--22-22-12", # KIA STINGER GT2 2018
+  "192a598e34926b1e|2019-04-04--13-27-39", # JEEP GRAND CHEROKEE 2019
+  "34a84d2b765df688|2018-08-28--12-41-00", # HONDA PILOT 2019 ELITE
+  "b0c9d2329ad1606b|2019-01-06--10-11-23", # CHRYSLER PACIFICA HYBRID 2017
+  "31390e3eb6f7c29a|2019-01-23--08-56-00", # KIA OPTIMA SX 2019
+  "fd10b9a107bb2e49|2018-07-24--16-32-42", # TOYOTA C-HR 2018
+  "9f7a7e50a51fb9db|2019-01-17--18-34-21", # JEEP GRAND CHEROKEE V6 2018
+  "aadda448b49c99ad|2018-10-25--17-16-22", # CHEVROLET MALIBU PREMIER 2017
+  "362d23d4d5bea2fa|2018-09-02--17-03-55", # TOYOTA HIGHLANDER HYBRID 2018
+  "1582e1dc57175194|2018-08-15--07-46-07", # HONDA ACCORD 2018 LX 1.5T
+  "fd10b9a107bb2e49|2018-07-24--20-32-08", # TOYOTA C-HR 2018
+  "265007255e794bce|2018-11-24--22-08-31", # CADILLAC ATS Premium Performance 2018
+  "53ac3251e03f95d7|2019-01-10--13-43-32", # HYUNDAI ELANTRA LIMITED ULTIMATE 2017
+  "21aa231dee2a68bd|2018-01-30--04-54-41", # HONDA ODYSSEY 2018 EX-L
+  "900ad17e536c3dc7|2018-04-12--22-02-36", # HONDA RIDGELINE 2017 BLACK EDITION
+  "975b26878285314d|2018-12-25--14-42-13", # CHRYSLER PACIFICA HYBRID 2018
+  "8ae193ceb56a0efe|2018-06-18--20-07-32", # TOYOTA RAV4 HYBRID 2017
+  "a893a80e5c5f72c8|2019-01-14--20-02-59", # HYUNDAI GENESIS 2018
+  "49c73650e65ff465|2018-11-19--16-58-04", # HOLDEN ASTRA RS-V BK 2017
+  "d2d8152227f7cb82|2018-07-25--13-40-56", # TOYOTA CAMRY 2018
+  "07cb8a788c31f645|2018-06-17--18-50-29", # mock
+  "c9d60e5e02c04c5c|2018-01-08--16-01-49", # HONDA CR-V 2016 TOURING
+  "1632088eda5e6c4d|2018-06-07--08-03-18", # HONDA CIVIC HATCHBACK 2017 SEDAN/COUPE 2019
+  "fbd011384db5e669|2018-07-26--20-51-48", # TOYOTA CAMRY HYBRID 2018
 ]
 
 if __name__ == "__main__":
 
+  # TODO: add routes for untested cars and fail test if we have an untested car
+  tested_cars = [keys["carFingerprint"] for route, keys in routes.items()]
+  for car_model in all_known_cars():
+    if car_model not in tested_cars:
+      print "***** WARNING: %s not tested *****" % car_model
+
   results = {}
   for route, checks in routes.items():
-
-    if route not in public_routes:
-      print "route not public", route
+    if route not in non_public_routes:
+      get_route_logs(route)
+    elif "UNLOGGER_PATH" not in os.environ:
       continue
-
-    get_route_logs(route)
 
     for _ in range(3):
       shutil.rmtree('/data/params')
@@ -420,7 +467,10 @@ if __name__ == "__main__":
 
       # Start unlogger
       print "Start unlogger"
-      unlogger_cmd = [os.path.join(BASEDIR, 'tools/replay/unlogger.py'), '%s' % route, '/tmp', '--disable', 'frame,plan,pathPlan,liveLongitudinalMpc,radarState,controlsState,liveTracks,liveMpc,sendcan,carState,carControl', '--no-interactive']
+      if route in non_public_routes:
+        unlogger_cmd = [os.path.join(BASEDIR, os.environ['UNLOGGER_PATH']), '%s' % route, '--disable', 'frame,plan,pathPlan,liveLongitudinalMpc,radarState,controlsState,liveTracks,liveMpc,sendcan,carState,carControl', '--no-interactive']
+      else:
+        unlogger_cmd = [os.path.join(BASEDIR, 'tools/replay/unlogger.py'), '%s' % route, '/tmp', '--disable', 'frame,plan,pathPlan,liveLongitudinalMpc,radarState,controlsState,liveTracks,liveMpc,sendcan,carState,carControl', '--no-interactive']
       unlogger = subprocess.Popen(unlogger_cmd, preexec_fn=os.setsid)
 
       print "Check sockets"