+++ RELEASES.md
@@ -1,3 +1,13 @@
+Version 0.6.4 (2019-09-08)
+========================
+ * Forward stock AEB for Honda Nidec
+ * Improve lane centering on banked roads
+ * Always-on forward collision warning
+ * Always-on driver monitoring, except for right hand drive countries
+ * Driver monitoring learns the user's normal driving position
+ * Honda Fit support thanks to energee!
+ * Lexus IS support
+
 Version 0.6.3 (2019-08-12)
 ========================
  * Alert sounds from EON: requires NEOS update